poc(KlaviyoCore): move retry, error-classification types from KlaviyoSwift
Moves the following generic networking concerns from KlaviyoSwift to KlaviyoCore:
- RetryState: retry posture enum used by the request queue
- NetworkingConstants.maxBackoff: 3-minute back-off ceiling
- KlaviyoEndpoint.maxRetries: per-endpoint retry cap
- InvalidField + parseError(): Klaviyo API error-response classification
- ErrorResponse/ErrorDetail/ErrorSource: internal decoding models

handleRequestError() stays in KlaviyoSwift as it returns KlaviyoAction.
