+++ Sources/KlaviyoSwift/StateManagement/StateManagement.swift
@@ -26,28 +26,6 @@ enum StateManagementConstants {
     static let initialAttempt = 1
 }
 
-/// Describes how the state machine should handle retrying a request after a failure.
-enum RetryState: Equatable {
-    /// Indicates that the request should be retried immediately (subject to
-    /// the regular flush cadence).
-    ///
-    /// - Parameter currentCount: The attempt number for the *current* request.
-    ///   The value should start at `1` for the very first send and is incremented each
-    ///   time a transient failure (such as a network error) occurs.
-    case retry(_ currentCount: Int)
-
-    /// Indicates that the request should be retried after waiting for a
-    /// server-specified back-off interval. This path is typically triggered by
-    /// an HTTP 429 "Too Many Requests" response that includes a `Retry-After`
-    /// header.
-    ///
-    /// - Parameters:
-    ///   - requestCount: The number of attempts made for this specific request.
-    ///   - totalRetryCount: The total number of attempts made for this request across all retry strategies.
-    ///   - currentBackoff: The remaining time in seconds to wait before the next retry attempt.
-    case retryWithBackoff(requestCount: Int, totalRetryCount: Int, currentBackoff: Int)
-}
-
 enum KlaviyoAction: Equatable {
     /// Sets the API key to state. If the state is already initialized then the push token is moved over to the company with the API key provided in this action.
     /// Loads the state from disk and carries over existing items from the queue. This emits `completeInitialization` at the end with the state loaded from disk.
